+++ src/main/java/ListGPT/Event.java
@@ -4,8 +4,8 @@ import java.time.DateTimeException;
 import java.time.LocalDate;
 
 public class Event extends Task {
-    LocalDate startDate;
-    LocalDate endDate;
+    private LocalDate startDate;
+    private LocalDate endDate;
 
     public Event(String name, String startDate, String endDate) throws DateTimeException {
         super(name);
@@ -16,9 +16,13 @@ public class Event extends Task {
             throw new DateTimeException("Invalid date format.");
         }
 
-        this.startDate = LocalDate.of(Integer.parseInt(startParams[0]), Integer.parseInt(startParams[1]),
+        this.startDate = LocalDate.of(
+                Integer.parseInt(startParams[0]),
+                Integer.parseInt(startParams[1]),
                 Integer.parseInt(startParams[2]));
-        this.endDate = LocalDate.of(Integer.parseInt(endParams[0]), Integer.parseInt(endParams[1]),
+        this.endDate = LocalDate.of(
+                Integer.parseInt(endParams[0]),
+                Integer.parseInt(endParams[1]),
                 Integer.parseInt(endParams[2]));
     }
 
@@ -34,6 +38,4 @@ public class Event extends Task {
     public String toString() {
         return "[E]" + super.toString() + " (from: " + this.getFrom() + " to:" + this.getTo() + ")";
     }
-
-
 }
