+++ src/main/java/ListGPT/ToDo.java
@@ -1,8 +1,8 @@
 package ListGPT;
 
 public class ToDo extends Task {
-    public ToDo(String name) {
-        super(name);
+    public ToDo(String description) {
+        super(description);
     }
 
     @Override
