+++ src/main/java/ListGPT/UI.java
@@ -3,34 +3,34 @@ package ListGPT;
 import java.util.Scanner;
 
 public class UI {
-    private final String name = "ListGPT";
-    private final String linebreak = "__________________________________";
+    private final String appName = "ListGPT";
+    private final String lineBreak = "__________________________________";
     private TaskList list;
 
     public UI(TaskList list) {
         this.list = list;
     }
 
-    public String message(String text) {
-        return "   " + linebreak + "\n" + text + "\n   " + linebreak + "\n";
+    public String formatMessage(String text) {
+        return "   " + lineBreak + "\n" + text + "\n   " + lineBreak + "\n";
     }
 
     public void run() {
-        Scanner sc = new Scanner(System.in);
-        System.out.println(message("   Hello! I'm " + name + "!\n   What can I do for you?"));
+        Scanner scanner = new Scanner(System.in);
+        System.out.println(formatMessage("   Hello! I'm " + appName + "!\n   What can I do for you?"));
 
         while (true) {
-            String input = sc.nextLine();
+            String input = scanner.nextLine();
             if (input.equals("bye")) {
-                this.bye();
+                this.sayBye();
                 return;
             }
             String response = Parser.parse(input, list);
-            System.out.print(message(response) + "\n");
+            System.out.print(formatMessage(response) + "\n");
         }
     }
 
-    private void bye() {
-        System.out.println(message("   Bye. Hope to see you again soon!"));
+    private void sayBye() {
+        System.out.println(formatMessage("   Bye. Hope to see you again soon!"));
     }
 }
