--- src/main/java/listgpt/InvalidTaskException.java
@@ -1,7 +1,0 @@
-package listgpt;
-
-public class InvalidTaskException extends RuntimeException {
-    public InvalidTaskException(String message) {
-        super(message);
-    }
-}
