--- src/main/java/listgpt/ListGPT.txt
@@ -1,3 +1,0 @@
-T | 1 | read book
-D | 0 | return book | June 6th
-E | 0 | project meeting | Aug 6th 2-4pm
