--- src/main/java/listgpt/ToDo.java
@@ -1,12 +1,0 @@
-package listgpt;
-
-public class ToDo extends Task {
-    public ToDo(String description) {
-        super(description);
-    }
-
-    @Override
-    public String toString() {
-        return "[T]" + super.toString();
-    }
-}
