+++ src/test/java/ListGPT/DeadlineTest.java
@@ -11,13 +11,13 @@ class DeadlineTest {
     @Test
     void constructor_validIsoDate_setsDueDate() {
         Deadline d = new Deadline("Submit report", "2025-08-10");
-        assertEquals("2025-08-10", d.getBy());
+        assertEquals("2025-08-10", d.getDueDate());
     }
 
     @Test
-    void getBy_returnsIsoString() {
+    void getDueDate_returnsIsoString() {
         Deadline d = new Deadline("Pay bills", "2030-01-01");
-        assertEquals("2030-01-01", d.getBy());
+        assertEquals("2030-01-01", d.getDueDate());
     }
 
     @Test
